Extract more shared helpers
